+++ crawl/src/main/java/org/newshabit/app/crawl/application/port/CrawlUseCase.java
@@ -3,6 +3,6 @@ package org.newshabit.app.crawl.application.port;
 import java.util.List;
 import org.newshabit.app.common.domain.model.CrawledNews;
 
-public interface CrawlInputPort {
+public interface CrawlUseCase {
 	List<CrawledNews> crawlNews();
 }
