+++ crawl/src/main/java/org/newshabit/app/crawl/application/port/MessageUseCase.java
@@ -3,6 +3,6 @@ package org.newshabit.app.crawl.application.port;
 import java.util.List;
 import org.newshabit.app.common.domain.model.CrawledNews;
 
-public interface MessageInputPort {
+public interface MessageUseCase {
 	void publishCrawledNews(List<CrawledNews> newsList);
 }
