+++ crawl/src/main/java/org/newshabit/app/crawl/application/schedule/CrawlingPublisherScheduler.java
@@ -5,8 +5,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.newshabit.app.common.domain.model.CrawledNews;
-import org.newshabit.app.crawl.application.port.CrawlInputPort;
-import org.newshabit.app.crawl.application.port.MessageInputPort;
+import org.newshabit.app.crawl.application.port.CrawlUseCase;
+import org.newshabit.app.crawl.application.port.MessageUseCase;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -14,15 +14,15 @@ import org.springframework.stereotype.Component;
 @Log4j2
 @Component
 public class CrawlingPublisherScheduler {
-	private final CrawlInputPort crawlInputPort;
-	private final MessageInputPort messageInputPort;
+	private final CrawlUseCase crawlUseCase;
+	private final MessageUseCase messageUseCase;
 
 	@Scheduled(cron = "0 0 * * * *")
 	public void crawlNewsAndProduce() {
 		try {
 			log.info("CrawlingPublisherScheduler Started: {}", LocalDateTime.now());
-			List<CrawledNews> crawledNews = crawlInputPort.crawlNews();
-			messageInputPort.publishCrawledNews(crawledNews);
+			List<CrawledNews> crawledNews = crawlUseCase.crawlNews();
+			messageUseCase.publishCrawledNews(crawledNews);
 			log.info("CrawlingPublisherScheduler Ended: {}: count: {}", LocalDateTime.now(), crawledNews.size());
 		} catch (Exception e) {
 			log.error(e);
