+++ crawl/src/main/java/org/newshabit/app/crawl/application/service/MessageService.java
@@ -1,16 +1,16 @@
-package org.newshabit.app.crawl.application.usecase;
+package org.newshabit.app.crawl.application.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.newshabit.app.common.domain.model.CrawledNews;
-import org.newshabit.app.crawl.application.port.MessageInputPort;
+import org.newshabit.app.crawl.application.port.MessageUseCase;
 import org.newshabit.app.crawl.application.port.MessageOutputPort;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MessageUseCase implements MessageInputPort {
+public class MessageService implements MessageUseCase {
 	private final MessageOutputPort<CrawledNews> messageOutputPort;
 
 	@Value("${app.kafka.crawl.topic}")
