+++ pubsub/src/main/java/org/newshabit/app/pubsub/application/service/RefineNewsService.java
@@ -1,4 +1,4 @@
-package org.newshabit.app.pubsub.application.usecase;
+package org.newshabit.app.pubsub.application.service;
 
 import java.io.IOException;
 import java.util.Optional;
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.newshabit.app.common.domain.model.RefinedNews;
 import org.newshabit.app.pubsub.application.port.AiOutputPort;
-import org.newshabit.app.pubsub.application.port.RefineNewsInputPort;
+import org.newshabit.app.pubsub.application.port.RefineNewsUseCase;
 import org.newshabit.app.pubsub.application.port.RefinedNewsRepositoryOutputPort;
 import org.newshabit.app.common.domain.entity.RefinedNewsEntity;
 import org.newshabit.app.common.domain.model.CrawledNews;
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class RefineNewsUseCase implements RefineNewsInputPort {
+public class RefineNewsService implements RefineNewsUseCase {
 	private final AiOutputPort aiOutputPort;
 	private final RefinedNewsRepositoryOutputPort refinedNewsRepositoryOutputPort;
