+++ pubsub/src/main/java/org/newshabit/app/pubsub/infrastructure/config/MessageConfig.java
@@ -5,7 +5,7 @@ import java.util.function.Function;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.newshabit.app.common.domain.model.RefinedNews;
-import org.newshabit.app.pubsub.application.port.RefineNewsInputPort;
+import org.newshabit.app.pubsub.application.port.RefineNewsUseCase;
 import org.newshabit.app.common.domain.model.CrawledNews;
 import org.newshabit.app.common.util.SleepUtil;
 import org.springframework.context.annotation.Bean;
@@ -17,14 +17,14 @@ import org.springframework.messaging.Message;
 @RequiredArgsConstructor
 @Slf4j
 public class MessageConfig {
-	private final RefineNewsInputPort refineNewsInputPort;
+	private final RefineNewsUseCase refineNewsUseCase;
 
 	@Bean
 	public Function<Message<CrawledNews>, Message<RefinedNews>> refineNews() throws RuntimeException {
 		return message -> {
 			try {
 				SleepUtil.randomSleep(1000, 3000);
-				return refineNewsInputPort.refineCrawledNews(message.getPayload())
+				return refineNewsUseCase.refineCrawledNews(message.getPayload())
 					.map(refinedNews -> MessageBuilder.withPayload(refinedNews).build())
 					.orElse(null);
 			} catch (Exception e) {
@@ -36,6 +36,6 @@ public class MessageConfig {
 
 	@Bean
 	public Consumer<Message<RefinedNews>> sinkRefinedNews() {
-		return message -> refineNewsInputPort.sinkRefinedNews(RefinedNews.toRefinedNewsEntity(message.getPayload()));
+		return message -> refineNewsUseCase.sinkRefinedNews(RefinedNews.toRefinedNewsEntity(message.getPayload()));
 	}
 }
